js/xpconnect/tests/chrome/test_bug601803.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=500931
-->
<window title="Mozilla Bug 601803"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=601803"
     target="_blank">Mozilla Bug 601803</a>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[

  /** Test for Bug 601803 **/

  function go() {
    var ifr = document.getElementById('ifr');
    var elem = document.createElementNS("http://www.w3.org/1999/xhtml","html:div");
    elem.appendChild(document.createTextNode("hello, world"));
    elem.expando = 42;
    ifr.contentDocument.body.appendChild(elem);
    is(elem.wrappedJSObject.expando, 42, "expando is preserved");
    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();

  ]]></script>
  <iframe type="content" src="about:blank" onload="go()" id="ifr" />
  </body>
</window>
